Tree: move delta
