+++ packages/dds/tree/src/changeset/index.ts
@@ -7,10 +7,6 @@
  * This module contains the changeset format and related operations.
  */
 
-// Split this up into seperate import and export for compatibility with API-Extractor.
-import * as Delta from "./delta";
-export { Delta };
-
 export * from "./format";
 export * from "./toDelta";
 export * from "./visit";
