+++ packages/dds/tree/src/feature-libraries/forestIndex.ts
@@ -17,8 +17,7 @@ import {
 } from "../forest";
 import { Index, SummaryElement } from "../shared-tree-core";
 import { cachedValue, ICachedValue, recordDependency } from "../dependency-tracking";
-import { Delta } from "../changeset";
-import { PlaceholderTree } from "../tree";
+import { PlaceholderTree, Delta } from "../tree";
 import { applyDeltaToForest } from "../transaction";
 import { placeholderTreeFromCursor, TextCursor } from "./treeTextCursor";
