+++ packages/dds/tree/src/tree/delta.ts
@@ -4,8 +4,9 @@
  */
 
 import { unreachableCase } from "@fluidframework/common-utils";
-import { FieldKey, JsonableTree, Value } from "../tree";
 import { Brand, Opaque } from "../util";
+import { FieldKey, Value } from "./types";
+import { JsonableTree } from "./treeTextFormat";
 
 /**
  * This format describes changes that must be applied to a document tree in order to update it.
