+++ packages/dds/tree/src/tree/index.ts
@@ -18,3 +18,7 @@ export {
 export * from "./pathTree";
 export * from "./anchorSet";
 export * from "./treeTextFormat";
+
+// Split this up into separate import and export for compatibility with API-Extractor.
+import * as Delta from "./delta";
+export { Delta };
